SelfAttn/requirements.txt
numpy
tensorflow
pyyaml
scikit-learn